add models and db to have timestamps.
