+++ models/User.model.js
@@ -2,43 +2,46 @@ const mongoose = require("mongoose");
 const { isEmail } = require("validator");
 const bcrypt = require("bcrypt");
 
-const userSchema = new mongoose.Schema({
-  email: {
-    type: String,
-    required: [true, "Please enter an email"],
-    unique: true,
-    lowercase: true,
-    validate: [isEmail, "Please enter a valid email"],
+const userSchema = new mongoose.Schema(
+  {
+    email: {
+      type: String,
+      required: [true, "Please enter an email"],
+      unique: true,
+      lowercase: true,
+      validate: [isEmail, "Please enter a valid email"],
+    },
+    password: {
+      type: String,
+      required: [true, "Please enter a password"],
+      minlength: [6, "Minimum password length is 6 characters"],
+    },
+    fullname: { type: String, require: [true, "Please enter your full names"] },
+    birthday: {
+      type: Date,
+    },
+    height: {
+      type: Number,
+      default: 0,
+    },
+    weight: {
+      type: Number,
+      default: 0,
+    },
+    gender: {
+      type: String,
+      enum: ["M", "F"],
+    },
+    bmr: {
+      type: Number,
+    },
+    hasBmr: {
+      type: Boolean,
+      default: false,
+    },
   },
-  password: {
-    type: String,
-    required: [true, "Please enter a password"],
-    minlength: [6, "Minimum password length is 6 characters"],
-  },
-  fullname: { type: String, require: [true, "Please enter your full names"] },
-  birthday: {
-    type: Date,
-  },
-  height: {
-    type: Number,
-    default: 0,
-  },
-  weight: {
-    type: Number,
-    default: 0,
-  },
-  gender: {
-    type: String,
-    enum: ["M", "F"],
-  },
-  bmr: {
-    type: Number,
-  },
-  hasBmr:{
-    type: Boolean,
-    default:false
-  }
-});
+  { timestamps: true }
+);
 
 // fire a function before doc saved to db
 userSchema.pre("save", async function (next) {
